// ui/lib/api/schemas/textWorkflow.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 */
import type { FontWorkflowTrace } from './fontWorkflowTrace';
import type { NodeId } from './nodeId';
import type { RepairWorkflowTrace } from './repairWorkflowTrace';
import type { TextResultMode } from './textResultMode';
import type { TextSelection } from './textSelection';
import type { TextWorkflowMode } from './textWorkflowMode';
import type { WorkflowStatus } from './workflowStatus';

export interface TextWorkflow {
  fontTrace?: null | FontWorkflowTrace;
  letteringStatus?: WorkflowStatus;
  modes?: TextWorkflowMode[];
  repairLayer?: null | NodeId;
  repairStatus?: WorkflowStatus;
  repairTrace?: null | RepairWorkflowTrace;
  resultMode?: TextResultMode;
  selection?: null | TextSelection;
}